Adds a new component  src/components/findJobsPage/FindRemoteJobs.tsx, Imports the new component to src/app/(root)/find-job/page.tsx
